bug: feign请求字段返回先转换成String再进行类型转换 #892
(cherry picked from commit ec4614784cb5bb0b8b2149dfec094cd53df93c37)

+++ src/backend/oci/biz-oci/src/main/kotlin/com/tencent/bkrepo/oci/service/impl/OciBlobServiceImpl.kt
@@ -142,7 +142,7 @@ class OciBlobServiceImpl(
             val nodeCreateRequest = ObjectBuildUtils.buildNodeCreateRequest(
                 projectId = projectId,
                 repoName = repoName,
-                size = nodeProperty.size!!.toLong(),
+                size = nodeProperty.size!!,
                 sha256 = ociDigest.hex,
                 fullPath = OciLocationUtils.buildDigestBlobsPath(packageName, ociDigest),
                 md5 = nodeProperty.md5!!
